requirements.txt
sqlalchemy
pymysql
streamlit
pandas
python-dotenv
workzeug